# docs/swagger.yaml
basePath: /
definitions:
  db.Task:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
  db.TaskPost:
    properties:
      description:
        type: string
      title:
        type: string
    type: object
  db.TaskPut:
    properties:
      description:
        type: string
      status:
        type: string
      title:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
host: localhost:8000
info:
  contact: {}
  description: This is a sample server celler server.
  title: CrudApp API
  version: "1.0"
paths:
  /tasks:
    get:
      consumes:
      - application/json
      description: Get all tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/db.Task'
            type: array
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method not allowed
      summary: GET
      tags:
      - tasks
    post:
      consumes:
      - application/json
      description: Add a new task to the list
      parameters:
      - description: Task Info
        in: body
        name: Task
        required: true
        schema:
          $ref: '#/definitions/db.TaskPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "500":
          description: Internal Server Error
      summary: POST
      tags:
      - tasks
  /tasks/{id}:
    delete:
      consumes:
      - application/json
      description: Delete task by ID
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
        "404":
          description: Not Found
        "405":
          description: Method not allowed
      summary: DELETE
      tags:
      - tasks
    put:
      consumes:
      - application/json
      description: Update an existing task
      parameters:
      - description: Id of a task.
        in: path
        name: id
        required: true
        type: integer
      - description: Task Info
        in: body
        name: Task
        required: true
        schema:
          $ref: '#/definitions/db.TaskPut'
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
      summary: PUT
      tags:
      - tasks
swagger: "2.0"